\name{CESdata}
\alias{CESdata}
\docType{data}
\title{Data for testing CES production function}
\description{Data for estimation and testing CES production function: q-output, l-labor, k-capital}
\usage{CESdata}
\format{
  A data frame with 25 observations on the following 3 variables.
  \describe{
    \item{\code{k}}{capital}
    \item{\code{l}}{labor}
    \item{\code{q}}{output}
     }
}
\source{EViews, coef_test.wf1}
\examples{
attach(CESdata)}

